Update RecurringScheduleTest and move TaskDoneStatusPredicateTest into predicate folder
