+++ src/test/java/seedu/address/model/task/predicates/TaskDoneStatusPredicateTest.java
@@ -1,4 +1,4 @@
-package seedu.address.model.task;
+package seedu.address.model.task.predicates;
 
 import static org.junit.jupiter.api.Assertions.*;
